+++ lib/Crypto/Hash/MD2.py
@@ -75,8 +75,8 @@ class MD2Hash(object):
     digest_size = 16
     #: The internal block size of the hash algorithm in bytes.
     block_size = 64
-
-    name = "md2"
+    #: ASN.1 Object ID
+    oid = "1.2.840.113549.2.2"
 
     def __init__(self, data=None):
         state = VoidPointer()