+++ lib/Crypto/SelfTest/Signature/__init__.py
@@ -28,7 +28,7 @@ import os
 
 def get_tests(config={}):
     tests = []
-    import test_pkcs1_15; tests += test_pkcs1_15.get_tests(config=config)
+    import test_pkcs1_v1_5; tests += test_pkcs1_v1_5.get_tests(config=config)
     import test_pkcs1_pss; tests += test_pkcs1_pss.get_tests(config=config)
     import test_dss; tests += test_dss.get_tests(config=config)
     return tests