+++ lib/Crypto/Signature/__init__.py
@@ -35,7 +35,6 @@ A collection of standardized protocols to carry out digital signatures.
 :undocumented: __revision__, __package__
 """
 
-__all__ = ['PKCS1_v1_5', 'PKCS1_PSS', 'DSS']
-__revision__ = "$Id$"
+__all__ = ['PKCS1_v1_5', 'PKCS1_PSS', 'DSS', 'pkcs1_v1_5']